fourchan: handle unsafe redirects
For whatever reason Fourchan started to use HTTP URLs when redirecting to some special pages (e.g. redirect to "not found" page if a thread is not found) and if the original URL was HTTPS then the client considers this redirect as unsafe and throws exception. That leads to "Unsafe redirect" toast message. HttpRequestUnsafeRedirectHandler class checks schemes of all redirects and if it finds "unsafe" redirects it replaces HTTP scheme with HTTPS and signals to the client to process the redirect.

Also in the original code some HTTP request were called with RedirectHandler.STRICT. I don't know how important is this so i delegated the redirect to this handler in the places when it was used in the original code.