fourchan: fix thread refresh when there are few posts in the thread
Recently Fourchan stopped provide a tail (most recent messages) for a thread with a few posts in the thread, now in this case it returns "404 not found" error page. I added ThreadsWithTailCache class that keeps numbers of threads that has enough messages to use tail request. These numbers are added when parsing a post in FourchanModelMapper. If a thread is long enough the first post will have "tail_size" property in JSON response. FourchanChanPerformer.onReadPosts now checks if the thread's number in the ThreadsWithTailCache before performing tail request.